protocol tc

+++ src/main/kotlin/com/duster/common/CommonMessageService.kt
@@ -99,7 +99,7 @@ val publisher = CommonPublisher
     {
         if (consumerMessageInDto.deliveryStatus!!.canReceiveFromConsumer) {
             messageStatusChangeHandler.updateDeliveryStatus(
-                consumerMessageInDto.id.toInt(),
+                consumerMessageInDto.id,
                 consumerMessageInDto.deliveryStatus!!,
                 Date(System.currentTimeMillis())
             )
@@ -116,7 +116,7 @@ val publisher = CommonPublisher
     }
 
 
-    fun getDeliveryStatusStatus(messageId: Int) : Optional<ProducerDeliveryStatusOutDto>
+    fun getDeliveryStatusStatus(messageId: Long) : Optional<ProducerDeliveryStatusOutDto>
     {
         return mainRepository.findDeliveredById(messageId)
             .map { ProducerDeliveryStatusOutDto(it) }
